Rawr.TankDK: Fixed the top two most expensive operations in calcs. Now runs at 1/6 original speed
git-tfs-id: [https://tfs.codeplex.com/tfs/TFS03]$/Rawr;C59319

+++ Rawr.DPSDK/DKAbilities/AbilityDK.FrostStrike.cs
@@ -76,7 +76,7 @@ namespace Rawr.DK
         }
 
         
-        public override int GetTotalDamage()
+        public override float GetTotalDamage()
         {
             if (CState.m_Spec == Rotation.Type.Frost)
                 return base.GetTotalDamage();
