+++ Rawr.DPSDK/DKAbilities/AbilityDK.HeartStrike.cs
@@ -56,7 +56,7 @@ namespace Rawr.DK
             }
         }
 
-        public override int GetTotalDamage()
+        public override float GetTotalDamage()
         {
             if (CState.m_Spec == Rotation.Type.Blood)
                 return base.GetTotalDamage();
